* commit 'c42de09a947fa511cc75a624c40b2d2816b8c0ee':
  AArch64: Use long for pointers in SurfaceTexture

+++ graphics/java/android/graphics/SurfaceTexture.java
@@ -69,9 +69,9 @@ public class SurfaceTexture {
     /**
      * These fields are used by native code, do not access or modify.
      */
-    private int mSurfaceTexture;
-    private int mBufferQueue;
-    private int mFrameAvailableListener;
+    private long mSurfaceTexture;
+    private long mBufferQueue;
+    private long mFrameAvailableListener;
 
     /**
      * Callback interface for being notified that a new stream frame is available.
